requirements.txt
flask
requests
flask_restful
pyyaml